+++ tests/FilterHttpTest.php
@@ -52,24 +52,44 @@ class FilterHttpTest extends TestCase
             ->assertJson(['posts' => [$this->postC->toArray()]]);
     }
 
-    public function testFilterWithLikeOperatorAndArrayValuesThrowsException()
+    public function testFilterWithLikeOperatorAndArrayValuesCanBeSearched()
     {
-        $this->getJson('/posts?search[]=lorem&search[]=ipsum')
-            ->assertJson(['exception' => 'InvalidArgumentException']);
+        $this->getJson('/posts?search[]=lorem, ipsum&search[]=ipsum')
+            ->assertJson(['posts' => [$this->postB->toArray(), $this->postC->toArray()]]);
+    }
+
+    public function testFilterWithRelatedModelsAttrCanBeMatched()
+    {
+        $this->getJson('/posts?comment=I am a comment,I am another comment')
+            ->assertJson(['posts' => [$this->postA->toArray(), $this->postB->toArray()]]);
+    }
+
+    public function testFilterWithNestedRelatedModelsAttrCanBeMatched()
+    {
+        $this->getJson('/posts?reply=I am a reply')
+            ->assertJson(['posts' => [$this->postA->toArray()]]);
     }
 
     protected function setUp(): void
     {
         parent::setUp();
+
         $this->postA = Post::create([
             'title' => 'lorem',
             'body'  => $this->faker->paragraph,
         ]);
+
+        $comment = $this->postA->comments()->create(['body' => 'I am a comment']);
+        $comment->replies()->create(['body' => 'I am a reply']);
+
         $this->postB = Post::create([
             'title'        => 'ipsum',
             'body'         => $this->faker->paragraph,
             'is_published' => 0,
         ]);
+
+        $this->postB->comments()->create(['body' => 'I am another comment']);
+
         $this->postC = Post::create([
             'title' => 'lorem, ipsum',
             'body'  => $this->faker->paragraph,